+++ src/routes/graphql/entities/user/user.type.ts
